inference_herbs_api/code/requirement.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.1.1
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==10.1.0
protobuf==4.24.4
psutil==5.9.6
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
Python-IO==0.3
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
unicorn==2.0.1.post1
urllib3==2.0.7
uvicorn==0.23.2
wandb==0.15.12
watchdog==3.0.0
zipp==3.17.0
